Made routing more scalable
