+++ src/main.js
@@ -4,7 +4,7 @@ import { createRouter, createWebHistory } from 'vue-router'
 import './style.css'
 import App from './App.vue'
 
-import { routes } from './router'
+import { routes } from './routing/router'
 
 const router = createRouter({
   history: createWebHistory(),
